pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email-service"
version = "2.0.0"
description = "Production-ready asynchronous email queue system for bookings and notifications"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Odiseo", email = "odiseo@odiseo.io"}
]
keywords = [
    "email",
    "async",
    "queue",
    "postgresql",
    "smtp",
    "jinja2",
    "pydantic",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Email",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "psycopg2-binary>=2.9.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "mypy>=1.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "isort>=5.12",
    "pre-commit>=3.0",
]

docs = [
    "sphinx>=6.0",
    "sphinx-rtd-theme>=1.2",
    "myst-parser>=1.0",
]

[project.urls]
Homepage = "https://github.com/odiseo/email-service"
Documentation = "https://github.com/odiseo/email-service/blob/main/README.md"
Repository = "https://github.com/odiseo/email-service.git"
Issues = "https://github.com/odiseo/email-service/issues"

[tool.setuptools]
packages = ["email_service"]

[tool.setuptools.package-data]
email_service = [
    "templates/*.html",
    "templates/*.txt",
    ".env.example",
]

[tool.black]
line-length = 88
target-version = ["py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # Directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true
skip_glob = ["*/migrations/*"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # Pyflakes
    "I",     # isort
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "UP",    # pyupgrade
]
ignore = [
    "E501",  # line too long
    "W505",  # line too long
]
exclude = [
    ".git",
    ".venv",
    "build",
    "dist",
    ".eggs",
    "*.egg-info",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["psycopg2.*", "jinja2.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["email_service"]
branch = true
omit = [
    "*/tests/*",
    "*/site-packages/*",
    ".venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
precision = 2

[tool.coverage.html]
directory = "htmlcov"
